move projects subscription logic to a separate hook
