+++ src/shared/layouts/CommonSidenavLayout/components/SidenavContent/hooks/useProjectsData.ts
@@ -1,13 +1,8 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router";
-import { CommonEvent, CommonEventEmitter } from "@/events";
-import { authentificated, selectUser } from "@/pages/Auth/store/selectors";
-import { CommonService, GovernanceService, Logger } from "@/services";
-import { GovernanceActions } from "@/shared/constants";
+import { authentificated } from "@/pages/Auth/store/selectors";
 import { useRoutesContext } from "@/shared/contexts";
-import { Common } from "@/shared/models";
-import { generateCirclesDataForCommonMember } from "@/shared/utils";
 import {
   commonLayoutActions,
   ProjectsStateItem,
@@ -24,6 +19,7 @@ import {
   getParentItemIds,
   Item,
 } from "../../../../SidenavLayout/components/SidenavContent/components";
+import { useProjectsSubscription } from "./useProjectsSubscription";
 
 interface Return {
   currentCommonId: string | null;
@@ -36,55 +32,6 @@ interface Return {
   itemIdWithNewProjectCreation: string;
   parentItemIds: string[];
 }
-
-const getProjectItemFromCommon = async (
-  common: Common,
-  userId?: string,
-  initialItem?: ProjectsStateItem,
-): Promise<ProjectsStateItem> => {
-  const baseItem: Omit<
-    ProjectsStateItem,
-    "hasMembership" | "hasPermissionToAddProject"
-  > = {
-    commonId: common.id,
-    image: common.image,
-    name: common.name,
-    directParent: common.directParent,
-  };
-
-  if (initialItem) {
-    return {
-      ...baseItem,
-      hasMembership: initialItem.hasMembership,
-      hasPermissionToAddProject: initialItem.hasPermissionToAddProject,
-    };
-  }
-  if (!userId) {
-    return {
-      ...baseItem,
-      hasMembership: false,
-      hasPermissionToAddProject: false,
-    };
-  }
-
-  const [governance, commonMember] = await Promise.all([
-    GovernanceService.getGovernanceByCommonId(common.id),
-    CommonService.getCommonMemberByUserId(common.id, userId),
-  ]);
-
-  return {
-    ...baseItem,
-    hasMembership: Boolean(commonMember),
-    hasPermissionToAddProject: Boolean(
-      governance &&
-        commonMember &&
-        generateCirclesDataForCommonMember(
-          governance.circles,
-          commonMember.circleIds,
-        ).allowedActions[GovernanceActions.CREATE_PROJECT],
-    ),
-  };
-};
 
 export const useProjectsData = (): Return => {
   const dispatch = useDispatch();
@@ -99,10 +46,6 @@ export const useProjectsData = (): Return => {
   const { projects, areProjectsLoading, areProjectsFetched } = useSelector(
     selectCommonLayoutProjectsState,
   );
-  const user = useSelector(selectUser());
-  const userId = user?.uid;
-  const [updatedItemsQueue, setUpdatedItemsQueue] = useState<Common[][]>([]);
-  const nextUpdatedItems = updatedItemsQueue[0];
   const currentCommon = commons.find(
     ({ commonId }) => commonId === currentCommonId,
   );
@@ -141,6 +84,12 @@ export const useProjectsData = (): Return => {
   const itemIdWithNewProjectCreation = getItemIdWithNewProjectCreationByPath(
     location.pathname,
   );
+  useProjectsSubscription({
+    activeItemId,
+    areProjectsFetched,
+    commons,
+    projects,
+  });
 
   useEffect(() => {
     if (isAuthenticated) {
@@ -183,70 +132,6 @@ export const useProjectsData = (): Return => {
     dispatch(commonLayoutActions.setCurrentCommonId(activeItemId));
     dispatch(commonLayoutActions.clearProjects());
   }, [activeItemId]);
-
-  useEffect(() => {
-    if (!areProjectsFetched) {
-      return;
-    }
-
-    const unsubscribe = CommonService.subscribeToSubCommons(
-      activeItemId,
-      (data) => {
-        const commons = data.reduce<Common[]>((acc, { common, isRemoved }) => {
-          if (!isRemoved) {
-            CommonEventEmitter.emit(CommonEvent.CommonUpdated, common);
-            return acc.concat(common);
-          }
-
-          return acc;
-        }, []);
-
-        if (commons.length !== 0) {
-          setUpdatedItemsQueue((currentItems) =>
-            currentItems.concat([commons]),
-          );
-        }
-      },
-    );
-
-    return unsubscribe;
-  }, [areProjectsFetched, activeItemId]);
-
-  useEffect(() => {
-    if (!nextUpdatedItems) {
-      return;
-    }
-    if (nextUpdatedItems.length === 0) {
-      setUpdatedItemsQueue((currentItems) => currentItems.slice(1));
-      return;
-    }
-
-    (async () => {
-      try {
-        const items = await Promise.all(
-          nextUpdatedItems.map(async (nextUpdatedItem) => {
-            const existingItem =
-              commons.find((item) => item.commonId === nextUpdatedItem.id) ||
-              projects.find((item) => item.commonId === nextUpdatedItem.id);
-
-            return await getProjectItemFromCommon(
-              nextUpdatedItem,
-              userId,
-              existingItem,
-            );
-          }),
-        );
-
-        items.forEach((item) => {
-          CommonEventEmitter.emit(CommonEvent.ProjectCreatedOrUpdated, item);
-        });
-      } catch (error) {
-        Logger.error(error);
-      } finally {
-        setUpdatedItemsQueue((currentItems) => currentItems.slice(1));
-      }
-    })();
-  }, [nextUpdatedItems]);
 
   return {
     currentCommonId,
